# backend/swagger.yaml
openapi: 3.0.0
info:
  title: WaysAhead Global API
  version: "1.0.0"
servers:
  - url: http://localhost:5000
paths:
  /api/login:
    post:
      summary: Authenticate and get JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
        '401':
          description: Invalid credentials
  /api/news:
    get:
      summary: Get news updates (protected)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of news articles
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    title:
                      type: string
                    date:
                      type: string
                    excerpt:
                      type: string
  /api/chatbot:
    post:
      summary: Chatbot endpoint powered by OpenAI
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                message:
                  type: string
      responses:
        '200':
          description: Chatbot reply
          content:
            application/json:
              schema:
                type: object
                properties:
                  reply:
                    type: string
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT